Flaky fix for logged out users
Now faking a profile in the state JSON when one doesn't exist already (user is not logged in). Fix still isn't perfect and you may need to reload page up to twice till everything works.

Bumped version to 1.2.1 .
Thank you especially to @iPrintScreen for a clear report of the problem.